% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_pi_R}
\alias{update_pi_R}
\title{Update pi parameter in scNPM model}
\usage{
update_pi_R(group_t, gam, K)
}
\arguments{
\item{group_t}{Cluster assignment vector}

\item{gam}{Prior parameter}

\item{K}{Number of clusters}
}
\value{
Dirichlet sampling result for pi
}
\description{
Update pi parameter in scNPM model
}
